src/test/java/com/faizal/didemo/controller/ConstructorInjectedControllerTest.java
package com.faizal.didemo.controller;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

import com.faizal.didemo.service.GreetingServiceImpl;

public class ConstructorInjectedControllerTest {

	private ConstructorInjectedController constructorInjectedController;

	@Before
	public void setUp() throws Exception {
		this.constructorInjectedController = new ConstructorInjectedController(new GreetingServiceImpl());
	}

	@Test
	public void testGreeting() {
		assertEquals("Hello - I was injected via the constructor!!!", constructorInjectedController.sayHello());
	}
}
